added Loggers, used Lombok and added Exception handling, changed database from h2 to mysql

+++ spring-boot-learning/src/main/java/com/springproject/springboot/learning/service/DepartmentServiceImpl.java
@@ -1,6 +1,7 @@
 package com.springproject.springboot.learning.service;
 
 import com.springproject.springboot.learning.entity.Department;
+import com.springproject.springboot.learning.error.DepartmentNotFoundException;
 import com.springproject.springboot.learning.repository.DepartmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -31,20 +32,32 @@ public class DepartmentServiceImpl implements DepartmentService{
     }
 
     @Override
-    public Department fetchDepartment(Long id) {
-        return departmentRepository.findById(id).get();
+    public Department fetchDepartment(Long id) throws DepartmentNotFoundException {
+        Optional<Department> department = departmentRepository.findById(id);
+        if(!department.isPresent()){
+            throw new DepartmentNotFoundException("Department with the id " + id + " is not available");
+        }
+        return department.get();
     }
 
     @Override
-    public String deleteDepartment(Long id) {
-        String dept = departmentRepository.findById(id).get().toString();
+    public String deleteDepartment(Long id) throws DepartmentNotFoundException {
+        Optional<Department> department = departmentRepository.findById(id);
+        if(!department.isPresent()){
+            throw new DepartmentNotFoundException("Department with the id "+  id + " is not present");
+        }
+        String dept = department.get().toString();
         departmentRepository.deleteById(id);
         return "deleted " + dept;
     }
 
     @Override
-    public Department updatedepartment(Long id, Department department) {
-        Department deptDB = departmentRepository.findById(id).get();
+    public Department updatedepartment(Long id, Department department) throws DepartmentNotFoundException {
+        Optional<Department> dept = departmentRepository.findById(id);
+        if(!dept.isPresent()){
+            throw new DepartmentNotFoundException("Department with the id "+  id + " is not present");
+        }
+        Department deptDB = dept.get();
 
         if(Objects.nonNull(department.getDepartmentName()) && !"".equalsIgnoreCase(department.getDepartmentName())){
             deptDB.setDepartmentName(department.getDepartmentName());
